+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Added `generate_user_configs_yaml()` to generate commented `user_settings.yaml` templates from existing settings managers.
+
 ## [3.2.0] - 2026-06-10
 
 ### Added